requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
pymongo==4.6.1
beautifulsoup4==4.12.2
nltk==3.8.1
numpy>=2.0.0
pandas>=2.0.0
requests==2.31.0
python-jwt==4.0.0
PyJWT==2.8.0
firebase-admin==5.0.0
gunicorn==20.1.0
eventlet==0.35.2
